+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/ChatClientTest.kt
@@ -58,6 +58,7 @@ internal class ChatClientTest {
             notifications = mock(),
             tokenManager = FakeTokenManager("")
         )
+            .apply { setUser(User(), "someToken") }
         result = mutableListOf()
     }