+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/channel/list/viewmodel/ChannelsViewModelBinding.kt
@@ -1,0 +1,24 @@
+@file:JvmName("ChannelsViewModelBinding")
+
+package io.getstream.chat.android.ui.channel.list.viewmodel
+
+import androidx.lifecycle.LifecycleOwner
+import com.getstream.sdk.chat.viewmodel.channels.ChannelsViewModel
+import io.getstream.chat.android.ui.channel.list.ChannelsView
+
+@JvmName("bind")
+public fun ChannelsViewModel.bindView(
+    view: ChannelsView,
+    lifecycle: LifecycleOwner
+) {
+    state.observe(lifecycle) { channelState ->
+        when (channelState) {
+            is ChannelsViewModel.State.Loading -> view.showLoadingView()
+
+            is ChannelsViewModel.State.Result -> {
+                view.setChannels(channelState.channels)
+                view.hideLoadingView()
+            }
+        }
+    }
+}