+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/utils/extensions/Any.kt
@@ -1,0 +1,8 @@
+package io.getstream.chat.android.ui.utils.extensions
+
+internal inline fun <reified AnyT> Any.cast() = this as AnyT
+
+internal inline fun <reified AnyT> Any.safeCast() = this as? AnyT
+
+internal val <AnyT> AnyT.exhaustive: AnyT
+    get() = this