+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/utils/extensions/Message.kt
@@ -1,0 +1,9 @@
+package io.getstream.chat.android.ui.utils.extensions
+
+import io.getstream.chat.android.client.models.Message
+import io.getstream.chat.android.client.models.name
+
+internal fun Message.getDisplayName() = when {
+    user.isCurrentUser() -> "You"
+    else -> user.name
+}