+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/controller/ChannelControllerImpl.kt
@@ -43,6 +43,7 @@ import io.getstream.chat.android.client.models.EventType
 import io.getstream.chat.android.client.models.Member
 import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.client.models.Reaction
+import io.getstream.chat.android.client.models.TypingEvent
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.client.utils.Result
 import io.getstream.chat.android.client.utils.SyncStatus
@@ -73,6 +74,7 @@ import java.util.Calendar
 import java.util.Date
 import java.util.concurrent.ConcurrentHashMap
 import kotlin.collections.set
+import kotlin.math.max
 
 private const val KEY_MESSAGE_ACTION = "image_action"
 private const val MESSAGE_ACTION_SHUFFLE = "shuffle"
@@ -133,13 +135,15 @@ internal class ChannelControllerImpl(
         .asLiveData()
 
     /** who is currently typing (current user is excluded from this) */
-    override val typing: LiveData<List<User>> = _typing
+    override val typing: LiveData<TypingEvent> = _typing
         .map {
-            it.values
+            val userList = it.values
                 .sortedBy(ChatEvent::createdAt)
                 .mapNotNull { event ->
                     (event as? TypingStartEvent)?.user ?: (event as? TypingStopEvent)?.user
                 }
+
+            TypingEvent(channelId, userList)
         }.asLiveData()
 
     /** how far every user in this channel has read */
@@ -841,7 +845,20 @@ internal class ChannelControllerImpl(
 
     private fun upsertMessages(messages: List<Message>) {
         val newMessages = parseMessages(messages)
+        updateLastMessageAtByNewMessages(newMessages.values)
         _messages.value = newMessages
+    }
+
+    private fun updateLastMessageAtByNewMessages(newMessages: Collection<Message>) {
+        if (newMessages.isEmpty()) {
+            return
+        }
+        val newLastMessageAt =
+            newMessages.mapNotNull { it.createdAt ?: it.createdLocallyAt }.maxOfOrNull(Date::getTime) ?: return
+        lastMessageAt.value = when (val currentLastMessageAt = lastMessageAt.value) {
+            null -> Date(newLastMessageAt)
+            else -> max(currentLastMessageAt.time, newLastMessageAt).let(::Date)
+        }
     }
 
     private fun upsertOldMessages(messages: List<Message>) {
@@ -1275,7 +1292,11 @@ internal class ChannelControllerImpl(
         return channel
     }
 
-    internal fun loadOlderThreadMessages(threadId: String, limit: Int, firstMessage: Message? = null): Result<List<Message>> {
+    internal fun loadOlderThreadMessages(
+        threadId: String,
+        limit: Int,
+        firstMessage: Message? = null
+    ): Result<List<Message>> {
         val result = if (firstMessage != null) {
             client.getRepliesMore(threadId, firstMessage.id, limit).execute()
         } else {