+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 ## stream-chat-android
 
 ## stream-chat-android-client
-- There is no more the requirement to wait set user before querying channels
+- It's no longer required to wait for `setUser` to finish before querying channels
 
 ## stream-chat-android-offline